storage: persist API server URL alongisde Account details (#14)
This commit adds the functionality that stores the acme-dns API URL alongisde of the rest of the account information. The account information is dependant of the API server in question, so at least for me it makes sense.

The new functionality allows the storage to be self sufficient for all the required data points. While the API server URL can often be deducted from the account URL, it's not guaranteed.

This commit also adds a test to ensure backwards compatibility with old storage format and to mitigate potential regressions in the backwards compatibility in the future.